fix: small fixes
